+++ package/api/src/controllers/metadatas/v1/mod.rs
@@ -2,7 +2,7 @@ use by_axum::{
     axum::{
         extract::{Path, Query, State},
         middleware,
-        routing::{post, get},
+        routing::post,
         Extension, Json, Router
     },
     log::root,
@@ -10,7 +10,7 @@ use by_axum::{
 use slog::o;
 
 use crate::{
-    common::CommonQueryResponse, middleware::auth::authorization_middleware, utils::error::ApiError,
+    common::CommonQueryResponse, middleware::auth::authorization_middleware,
 };
 
 use models::prelude::*;
@@ -21,11 +21,11 @@ pub struct MetadataControllerV1 {
 }
 
 impl MetadataControllerV1 {
-    pub fn router() -> Result<Router> {
+    pub fn router() -> Router {
         let log = root().new(o!("api-controller" => "MetadataControllerV1"));
         let ctrl = MetadataControllerV1 { log };
 
-        Ok(Router::new()
+        Router::new()
             .route("/", post(Self::act_metadata).get(Self::list_metadatas))
             .route(
                 "/:metadata_id",
@@ -33,7 +33,7 @@ impl MetadataControllerV1 {
             )
             .route("/upload", post(Self::upload_metadata))
             .with_state(ctrl)
-            .layer(middleware::from_fn(authorization_middleware)))
+            .layer(middleware::from_fn(authorization_middleware))
     }
 
     pub async fn act_metadata(
