+++ package/api/src/main.rs
@@ -7,20 +7,38 @@ mod common;
 mod controllers {
     pub mod auth {
         pub mod v1;
-    //     pub mod login {
-    //         pub mod v1;
-    //     }
-    //     pub mod reset {
-    //         pub mod v1;
-    //     }
-    //     pub mod signup {
-    //         pub mod v1;
-    //     }
+    }
+    pub mod verification {
+        pub mod v1;
+    }
+    pub mod members {
+        pub mod v1;
     }
     pub mod organizations {
         pub mod v1;
     }
     pub mod groups {
+        pub mod v1;
+    }
+    pub mod attributes {
+        pub mod v1;
+    }
+    pub mod metadatas {
+        pub mod v1;
+    }
+    pub mod search {
+        pub mod v1;
+    }
+    pub mod panels {
+        pub mod v1;
+    }
+    pub mod public_opinions {
+        pub mod v1;
+    }
+    pub mod public_surveys {
+        pub mod v1;
+    }
+    pub mod survey { // FIXME: deprecated
         pub mod v1;
     }
 }
@@ -51,13 +69,17 @@ async fn main() {
     );
 
     let app = by_axum::new()
-        // .nest("/auth/login/v1", controllers::auth::login::v1::LoginControllerV1::router())
-        // .nest("/auth/signup/v1", controllers::auth::signup::v1::SignupControllerV1::router())
-        // .nest("/auth/reset/v1", controllers::auth::reset::v1::ResetControllerV1::router())
         .nest("/auth/v1", controllers::auth::v1::AuthControllerV1::router())
+        .nest("/members/v1", controllers::members::v1::MemberControllerV1::router())
         .nest("/organizations/v1", controllers::organizations::v1::OrganizationControllerV1::router())
-        .nest("/groups/v1", controllers::groups::v1::GroupControllerV1::router()
-    );
+        .nest("/groups/v1", controllers::groups::v1::GroupControllerV1::router())
+        .nest("/attributes/v1", controllers::attributes::v1::AttributeControllerV1::router())
+        .nest("/metadatas/v1", controllers::metadatas::v1::MetadataControllerV1::router())
+        .nest("/search/v1", controllers::search::v1::SearchControllerV1::router())
+        .nest("/panels/v1", controllers::panels::v1::PanelControllerV1::router())
+        .nest("/public-opinions/v1", controllers::public_opinions::v1::PublicOpinionControllerV1::router())
+        .nest("/public-surveys/v1", controllers::public_surveys::v1::PublicSurveyControllerV1::router())
+        .nest("/survey/v1", controllers::survey::v1::AxumState::router()); // FIXME: deprecated
 
     #[cfg(feature = "reload")]
     {
