+++ package/api/src/utils/jwt/mod.rs
@@ -6,8 +6,10 @@ use jsonwebtoken::{
     Algorithm, DecodingKey, EncodingKey, Header, Validation,
 };
 use serde::{Deserialize, Serialize};
+use by_axum::aide;
+use schemars::JsonSchema;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, Debug, Clone, JsonSchema, aide::OperationIo)]
 pub struct Claims {
     pub id: String, //user id
     pub email: String,
